The C# port replays the golden corpus bit-exactly
9 runs, 98,970 ticks, 1,661 checkpoints, every world hash and every stats hash
identical:

    cd cs && dotnet run --project src/Scrapyard.Golden -- verify

Brings the last three pieces - Step.StepWorld (the stage order), Simulation (the
port of createWorld), and the Scrapyard.Golden runner that reads the corpus and
diffs it. The runner refuses to run on a formatVersion, hashAlgo or tickRate it
does not recognise: a golden master that misreads its own format and reports
success is worse than not having one.

Every system already had an adversarial fixture, and every fixture had been
proven to fail under injected faults. The port still had NINE defects and the
corpus found all nine, none of which was a mistranslated line:

FIVE UNINITIALISED FIELDS, the same mistake five times - Player.Level,
WeaponInstance's level/turretX/targetDense/ammo, DifficultyState's two ramps,
LevelUpState's offers and lastTaken, and RunStats.KilledByRank. Each is a field
whose "unset" value is not zero, and each was invisible because every fixture
wrote the fields it cared about before reading them.

TWO FIELDS THAT SHOULD NOT HAVE EXISTED - World.MaxEnemyRadius shadowed a
constant and World.PlayerRadius shadowed a resolved stat. Production never wrote
either, so no projectile could find a body more than its own radius away and no
enemy could reach the mech at all. CollisionTests passed throughout, because its
fixture set both: a test that supplies a value production forgets is not testing
that value, it is hiding it.

ONE NARROWED INTERFACE - EnemyAI and the spawn ring pushed bodies out of
ScrapPiles and did nothing for anything else, so on both lattice levels every
enemy walked through every wall. A leftover from before IScenery had PushOut.

ONE FLOAT32 SLIP, in the shape this project has now hit twice: `p.X[d] -
world.Enemies.X[e]` is float-minus-float, which C# evaluates in single precision
where JavaScript widens both reads. One ULP in a sheep's heading at tick 5,690.

The unit tests found the bugs inside functions; the corpus found the bugs between
them. cs/README.md records all nine and the three-step workflow that localised
them, because the next port of anything will make the same five mistakes.

+++ cs/src/Scrapyard.Core/Sheep.cs
@@ -272,8 +272,12 @@ public static class Sheep
                     for (int k = 0; k < n; k++)
                     {
                         int e = near[k];
-                        ax += p.X[d] - world.Enemies.X[e];
-                        ay += p.Y[d] - world.Enemies.Y[e];
+                        // (double) ON THE FIRST OPERAND. Both sides are float columns, and C#
+                        // would evaluate `float - float` in SINGLE precision and round before the
+                        // widening add - where JavaScript widens both reads and subtracts in
+                        // double. One ULP here, four thousand ticks later a different flock.
+                        ax += (double)p.X[d] - world.Enemies.X[e];
+                        ay += (double)p.Y[d] - world.Enemies.Y[e];
                     }
                     if (distP2 < AvoidRadius * AvoidRadius)
                     {
